Update README.md
Problem Statement: To understand what exactly a listener prefers listening to on radio, every detail is recorded online. This recorded information is used for recommending music that the listener is likely to enjoy and to come up with a focused marketing strategy that sends out advertisements for music that a listener may wish to buy. However, this results in wasting money on scarce advertising.